requirements.txt
Django==5.2.5
djangorestframework==3.16.1
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.4.0
psycopg2-binary==2.9.9